+++ client/next.config.mjs
@@ -12,7 +12,7 @@ const nextConfig = {
   },
   reactStrictMode: true,
   images: {
-    domains: ['localhost', '127.0.0.1'], 
+    domains: ['localhost', '127.0.0.1', 'squadrone.com.au'], 
   },
   // dumb fix for windows docker
   webpack: isWindowsDevContainer()
