+++ client/src/pages/guests-sponsors.tsx
@@ -1,24 +1,86 @@
-import { useState } from "react";
+import { Montserrat } from "next/font/google";
+import React, { useEffect, useState } from "react";
 
-import { usePings } from "@/hooks/pings";
+import { GuestSection, SponsorSection } from "@/components/SponsorsGuests";
 
-import { Button } from "../components/ui/button";
+type Partner = {
+  name: string;
+  logo: string;
+  link: string;
+};
 
-export default function Home() {
-  const [clicked, setClicked] = useState(false);
-  const { data, isLoading } = usePings({
-    enabled: clicked,
-  });
+// Type definition for guest data
+type Guest = {
+  name: string;
+  description: string;
+  link: string;
+  photo?: string;
+};
+
+const montserrat = Montserrat({
+  subsets: ["latin"],
+  weight: ["400", "600", "700"],
+});
+
+const SponsorsAndGuest = () => {
+  const [partners, setPartners] = useState<Partner[]>([]); // Holds sponsor data
+  const [guests, setGuests] = useState<Guest[]>([]); // Holds guest data
+  const [loading, setLoading] = useState(true); // Controls loading state
+
+  useEffect(() => {
+    // Define an async function to fetch data
+    const fetchData = async () => {
+      try {
+        const [partnerRes, guestRes] = await Promise.all([
+          fetch("http://localhost:8000/api/sponsor/"),
+          fetch("http://localhost:8000/api/sponsor/guests/"),
+        ]);
+        const partnerData = await partnerRes.json();
+        const guestData = await guestRes.json();
+
+        setPartners(partnerData);
+        setGuests(guestData);
+      } catch (error) {
+        console.error("Error fetching data:", error);
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchData();
+  }, []);
 
   return (
-    <>
-      <h1 className="title-large text-dark">Title Test</h1>
-      <Button onClick={() => setClicked(true)}>
-        {isLoading ? "Loading" : "Ping"}
-      </Button>
-      <p>
-        Response from server: <span>{data as string}</span>
-      </p>
-    </>
+    <div className="min-h-screen bg-background">
+      {/* Page title */}
+      <div className="pt-20 text-center">
+        <h1
+          className={`text-6xl font-semibold leading-none ${montserrat.className}`}
+        >
+          Guests & Sponsors
+        </h1>
+      </div>
+
+      {/* Special Guests Section */}
+      <GuestSection
+        guests={guests.map((guest) => ({
+          name: guest.name,
+          description: guest.description,
+          link: guest.link,
+          image: guest.photo, // mapped to 'image' for component
+        }))}
+      />
+
+      {/* Sponsors Section */}
+      <div className="mt-16">
+        {loading ? (
+          <p className="text-center text-gray-500">Loading sponsors...</p>
+        ) : (
+          <SponsorSection partners={partners} />
+        )}
+      </div>
+    </div>
   );
-}
+};
+
+export default SponsorsAndGuest;
